// frontend/src/modules/dashboard/types/dashboard.types.ts
export interface DashboardResumen {
  vehiculosParqueados: number;
  capacidadTotal: number;
  ingresosHoy: number;
  gastosHoy: number;
  balanceHoy: number;
  totalMovimientosHoy: number;
  cajaAbierta: boolean;
  metodosPago?: {
    efectivo: number;
    yape: number;
    plin: number;
    tarjeta: number;
  };
  ultimasEntradas?: {
    id: string;
    placa: string;
    tipoVehiculo: string;
    fechaEntrada: string | Date;
    momentoPago: string;
  }[];
  ultimasSalidas?: {
    id: string;
    placa: string;
    tipoVehiculo: string;
    fechaSalida: string | Date;
    totalPagar: number;
    metodoPago: string;
  }[];
}